requirements.txt
sqlalchemy==1.4.39
loguru==0.7.2
fastapi==0.96.1
pyjwt==2.4.0
uvicorn==0.30.0
pydantic==1.10.8
python-multipart==0.0.9